pyproject.toml
[tool.poetry]
name = "oppora-marketing"
version = "0.1.0"
description = "Automate marketing, social media and business growth."
authors = ["Hojayfa Rahman <hojayfarahman@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
dj-database-url = "^3.0.0"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
django-unfold = "^0.59.0"
pillow-avif-plugin = "^1.4.6"
pillow = "^11.1.0"
psycopg2 = "^2.9.10"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
requests = "^2.32.4"
langchain = "^0.3.27"
langchain-openai = "^0.3.28"
langchain-community = "^0.3.27"
langgraph = "^0.6.3"
beautifulsoup4 = "^4.13.4"
django-q2 = "^1.8.0"
cryptography = "^45.0.6"
django-safedelete = "^1.4.1"
pendulum = "^3.1.0"
dj-rest-auth = "^7.0.1"
django-allauth = "^65.11.0"
drf-api-logger = "^1.1.20"
django-simple-history = "^3.10.1"
django-import-export = "^4.3.9"
dj-stripe = "^2.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
pytest-cov = "^6.1.1"
factory-boy = "^3.3.3"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = ["--strict-markers", "--strict-config", "--verbose", "--tb=short"]
testpaths = ["common", "workspace"]
filterwarnings = [
	# dj-rest-auth referencing deprecated allauth settings at import-time
	'ignore:.*app_settings\.USERNAME_REQUIRED is deprecated.*:UserWarning:dj_rest_auth\.registration\.serializers',
	'ignore:.*app_settings\.EMAIL_REQUIRED is deprecated.*:UserWarning:dj_rest_auth\.registration\.serializers',
	'ignore:.*AUTHENTICATION_METHOD is deprecated.*:UserWarning:dj_rest_auth\.serializers',
]
